[#447] NEW: add example on hard-wait before command
Add example on how to «hard-wait» (sleep)
before any waiting command in Selene.

This is useful when you know that a certain amount of time
is needed for an element to appear or become available.

This commit introduces a new example that shows how to
use a custom context manager with `selene.support._wait.with_`
to add a `time.sleep` before Selene's waiting logic starts.